ajustes

+++ manifests/uninstall.pp
@@ -1,8 +1,9 @@
-define usvn::remove()
+define usvn::uninstall()
 {
 	include usvn::params
 	exec {'usvn::remove::install':
-		path    => '/bin',
+		path    => ['/bin','/usr/bin'],
 		command => "rm -Rf ${usvn::params::instdir}",
+		onlyif  => "test -d ${usvn::params::instdir}",
 	}
 }
